[legacy] cupy_portable_version/requirements.txt
h5py
numpy
scipy
pandas
ripser
scikit-learn
matplotlib
seaborn
imageio
pillow
pyvista
scikit-image
fastapi
uvicorn
websockets
watchdog
PyYAML==6.0.1
cupy-cuda12x